+++ test/testFoldr.js
@@ -7,27 +7,27 @@ define(['nu/stream'], function(stream){
     return {
         'module': "Foldr",
         'tests': [
-            ["Simple foldr",
+            ["Simple",
             function(){
                 assert.deepEqual(
                     stream.foldr(r, 3, stream.from([0, 1, 2])),
                     [0, [1, [2, 3]]]);
             }],
-            ["Single foldr",
+            ["Single",
             function(){
                 assert.deepEqual(
                     stream.foldr(r, 10, stream.from([0])),
                     [0, 10]);
             }],
-            ["foldr Index",
+            ["Index",
             function(){
                 assert.deepEqual(
-                    stream.foldr(function(p, c, i) {
-                        return p + c * i;
-                    }, 0, stream.from([0, 1, 2, 3])),
+                    stream.foldr(function(p, c) {
+                        return p + c[1] * c[0];
+                    }, 0, stream.indexed(stream.from([0, 1, 2, 3]))),
                     14);
             }],
-            ["Empty foldr",
+            ["Empty",
             function(){
                 assert.equal(
                     stream.foldr(r, 0, stream.end),