+++ test/testReduceRight.js
@@ -5,27 +5,27 @@ define(['nu/stream'], function(stream){
     }
     
     return {
-        'module': "reduceRight Tests",
+        'module': "reduceRight",
         'tests': [
-            ["Simple reduceRight",
+            ["Simple",
             function(){
                 assert.deepEqual(
                     stream.reduceRight(r, stream.from([0, 1, 2, 3])),
                     [0, [1, [2, 3]]]);
             }],
-            ["Single reduceRight",
+            ["Single",
             function(){
                 assert.deepEqual(
                     stream.reduceRight(r, stream.from([0])),
                     0);
             }],
-            ["reduceRight Index",
+            ["reduceRight",
             function(){
                 assert.deepEqual(
-                    stream.reduceRight(function(p, c, i) {
-                        return p + c * i;
-                    }, stream.from([0, 1, 2, 3])),
-                    8);
+                    stream.reduceRight(function(p, c) {
+                        return c[0];
+                    }, stream.indexed(stream.from([0, 1, 2, 3]))),
+                    0);
             }],
         ],
     };